Rego-based routing of Rego handlers + refactor (#1675)
This is a lot, and I'm sorry about that 😅 Should have pushed
something earlier, but was caught up in other things, and this
was mostly work done during travel and brief moments here and
there.

The main thing here is the new Rego-router, which handles the
requirements of individual routes, and uses a single prepared
query + dynamic policy composition for routing in Rego.

- Add new main entrypoint for Rego-based routes
- Rearrange files in LSP directory and use subpackages to avoid
  having everything in just one big dir/package
- Merge util package of roast with internal/util

There are now only 2 remaining completion providers written in
Go! These are a little more complex to port, but is of course
the obvious next step. This would allow us to get rid of the
Go-based completions manager entirely.

Signed-off-by: Anders Eknert <anders@eknert.com>