Fix documentLink issue when no user config provided
Pretty sure the documentLink issue Charlie found was only
the tip of the iceberg, as we did a lot of initialization
logic seemingly only when a custom user config was found..

Good to have that fixed!

Also, update references to docs.styra.com to now point at
their new OPA docs home.

UPDATE: It turned out to be the tip of an iceberg indeed,
as the root cause of the issue was shared query state between
goroutines. Unlikely to ever cause issues in "production", but
certainly so in parallel tests. Required quite some refactoring,
and I'm not too happy with the design... but the ugly parts will
all be elimintated once we get the last few components onboarded
on the Rego router, which is nice.

Signed-off-by: Anders Eknert <anders@eknert.com>