Builtin function completion in Rego (#1694)
Rewrite this completion provider from Go to Rego.

The only thing not ported from the Go implementation is rendering
of nested types. So for now we will only show e.g. `array` snd not
`array<string|number>` or whatever. This is doable in Rego, of course,
but it feels like it requires more effort than its worth for a first
implementation. Users who need this can always click to see the docs.

A *lot* of effort went into first understanding why `strings.render_template`
would fail so often, and once I did to try and work around that fact. We could
just have ignored builtin definitions without the attributes we need, but since
we had an integration test pulling in some ancient EOPA capabilities, and users
could provide any kind of caps themselves, I figured it was worth doing right.

This code also lays the foundation for rewriting the hover provider in Rego
too later.

As usual, a few things got refactored along the way.. but I did try to avoid
steering too far off the set course this time :)

Signed-off-by: Anders Eknert <anders@eknert.com>