Package ref import suggestions in Rego (#1698)
Rewrite the last Go completion provider in Rego!

The next step is to get tid of the now pointless Manager/Policy
Go completion handlers, and integrate Rego completions in the main
Rego router. This is a little more involved though, as the input
format currently used for completion providers is quite different
from what other Rego handlers are provided. Which in itself is a
good reason for why we'll want to fix it... but will follow up with
another PR for that later, so that this can be reviewed in isolation.

Signed-off-by: Anders Eknert <anders@eknert.com>