lsp: Implement selection ranges and linked editing range
Selection ranges provide smart selection of code that can expand and shrink based
on knowledge of the AST rather than simple text properties like hyphens or whitespace.
I didn't think this would be particularly important, but having it used it for some time
while working on this feature, this is definitely a feature that I now wouldn't want to
be without. While not likely to be used much outside of rules, ranges also expand in both
package declarations and imports.

Linked editing range OTOH was mostly implemented to try out the feature, and at this point
has not proven to be particularly useful, and the feature is therefore kept behind a new
`REGAL_EXPERIMENTAL` flag for now. More info about that in the docs included in this PR.

Also:
- Remove the "log only once" logic in the development bundle loader, as it's no longer flooding the logs
- Use shared schema for all `TextDocumentPosition` params handlers
- Document all environment variables available for development and debugging purposes

Signed-off-by: Anders Eknert <anders@eknert.com>