lsp: document highlighting supporting function args (#1718)
A function arg variable at requested position will now have the document
highlight handler return all references of that arg from the function
head/body for highlighting, making it easy to quickly identify the locations
where the arg is used.

Less interesting to users is the *much* improved REGAL_BUNDLE_PATH development
mode added with this PR. This is the result of a substantial side quest when
working on the highlighting feature, and finding the current implementation to
perform too poorly. Which was no surprise given how previously we'd re-prepare
the query on every call to `LoadedBundle`, which sometimes happens several
times per second!  The new implementation instead makes use of a watcher process,
and re-prepares the relevant LSP query only when a file in the bundle is changed
and that change is saved.

Also properly documented the dev mode feature as it now makes LSP development so
much more enjoyable.

Signed-off-by: Anders Eknert <anders@eknert.com>