Get FindClosestMatchingRoot working on windows (#1740)
* Get FindClosestMatchingRoot working on windows

There was an issue in this test where it handled the trimming of
prefixes incorrectly in a way that made the TestFixer test fail. This
updates the implementation to address this and adds some tests
to validate the windows behavior.

Signed-off-by: Charlie Egan <charlie_egan@apple.com>

* Address issue in TestFilterIgnoredPaths

normalize both filename and pathPrefix to forward slashes using
filepath.ToSlash() before the prefix trimming.

Signed-off-by: Charlie Egan <charlie_egan@apple.com>

* Normalise suffix to fix TestFindConfigRoots

Signed-off-by: Charlie Egan <charlie_egan@apple.com>

---------

Signed-off-by: Charlie Egan <charlie_egan@apple.com>